fix removes (#175)
related to
-
https://github.com/amtoine/nu-git-manager/actions/runs/7684621264/job/20942567622

## description
not sure why this did not fail before but it does now on the latest
`main` of Nushell :thinking:

## changelog
- the first commit is here to show the CI failing (well it did not run
at all, lol, but it fails on locally by using a _nightly_ build and the
latest revision of Nupm)
- the rest should fix the CI and the tests

+++ pkgs/nu-git-manager/tests/gm.nu
@@ -21,7 +21,9 @@ def run-with-env [code: closure, --prepare-cache] {
 
     with-env $TEST_ENV $code
 
-    rm --recursive --force --verbose $TEST_ENV_BASE
+    if ($TEST_ENV_BASE | path exists) {
+        rm --recursive --force --verbose $TEST_ENV_BASE
+    }
 }
 
 export def error-with-empty-store [] {
